Nnn read me
I think it will also be good for POS  it should then become  a more populated network with the help of dapp.nodes plug and play equipment hard-node. The ENS. will then have the 1st official dnssec node name network.nnn and will become the fastest most secure id verified  POS and that makes  it more cost effective to mint verify and document  and anyone who bothered to get a convenient. Eth name I would think would like to be a part of the 1st central linking sec approved 
Node blockchain name service network 
And then to find out that the hardnode will pay you for being plugged in and bucause of the shorter distance between the..Nnn s and the amount of them being used by the users that are making there secured data becomes the pieces that bind the chain with a stable hardnode are all  available  use the network all ready they they just need the ability and insentive to connect there data privately no actual network